+++ tests/int_bitbucket.rs
@@ -17,7 +17,7 @@ use kingfisher::{
             rules::RuleSpecifierArgs,
             scan::{ConfidenceLevel, ScanArgs},
         },
-        global::Mode,
+        global::{Mode, TlsMode},
         GlobalArgs,
     },
     findings_store::FindingsStore,
@@ -170,6 +170,7 @@ fn test_bitbucket_remote_scan() -> Result<()> {
         self_update: false,
         ignore_certs: false,
         user_agent_suffix: None,
+        tls_mode: TlsMode::Strict,
     };
 
     let datastore = Arc::new(Mutex::new(FindingsStore::new(clone_dir)));