+++ tests/int_gitlab.rs
@@ -18,7 +18,7 @@ use kingfisher::{
             rules::RuleSpecifierArgs,
             scan::{ConfidenceLevel, ScanArgs},
         },
-        global::Mode,
+        global::{Mode, TlsMode},
         GlobalArgs,
     },
     findings_store::FindingsStore,
@@ -175,6 +175,7 @@ fn test_gitlab_remote_scan() -> Result<()> {
         self_update: false,
         ignore_certs: false,
         user_agent_suffix: None,
+        tls_mode: TlsMode::Strict,
     };
 
     let datastore = Arc::new(Mutex::new(FindingsStore::new(clone_dir)));
@@ -338,6 +339,7 @@ fn test_gitlab_remote_scan_no_history() -> Result<()> {
         self_update: false,
         ignore_certs: false,
         user_agent_suffix: None,
+        tls_mode: TlsMode::Strict,
     };
 
     let datastore = Arc::new(Mutex::new(FindingsStore::new(clone_dir)));