+++ tests/int_slack.rs
@@ -14,7 +14,7 @@ use kingfisher::{
             rules::RuleSpecifierArgs,
             scan::{ConfidenceLevel, ScanArgs},
         },
-        global::Mode,
+        global::{Mode, TlsMode},
         GlobalArgs,
     },
     findings_store::FindingsStore,
@@ -308,6 +308,7 @@ async fn test_scan_slack_messages() -> Result<()> {
         progress: Mode::Never,
         ignore_certs: false,
         user_agent_suffix: None,
+        tls_mode: TlsMode::Strict,
     };
 
     let datastore = Arc::new(Mutex::new(FindingsStore::new(clone_dir)));